Lots of updates / fixes to the last setup
Wanted to make sure it was working properly, changed a few things.

Now the user only enters the directory. I also previously mentioned file permissions, that was actually in this commit.

Should be able to do the entire setup process without errors now. It should be obvious when you're doing something wrong as well.