#60 Enforce Unique Global Mentions Platform, Cleanup
Added a check to enforce unique platforms for global IDs. Might add the ability to use multiple IDs of the same platform for a single global ID in the future, but for now you can just register multiple global IDs.

Also did some cleanup / abstraction of the code to make it more readable.